+++ src/error.rs
@@ -30,9 +30,9 @@ use crate::{
     PeerId,
 };
 
-use multiaddr::Multiaddr;
-use multihash::{Multihash, MultihashGeneric};
+type Multihash = multihash::Multihash<64>;
 
+use multiaddr::Multiaddr;
 use std::io::{self, ErrorKind};
 
 // TODO: https://github.com/paritytech/litep2p/issues/204 clean up the overarching error.
@@ -426,8 +426,8 @@ pub enum DnsError {
     IpVersionMismatch,
 }
 
-impl From<MultihashGeneric<64>> for Error {
-    fn from(hash: MultihashGeneric<64>) -> Self {
+impl From<Multihash> for Error {
+    fn from(hash: Multihash) -> Self {
         Error::AddressError(AddressError::InvalidPeerId(hash))
     }
 }
@@ -498,8 +498,8 @@ impl From<ParseError> for Error {
     }
 }
 
-impl From<MultihashGeneric<64>> for AddressError {
-    fn from(hash: MultihashGeneric<64>) -> Self {
+impl From<Multihash> for AddressError {
+    fn from(hash: Multihash) -> Self {
         AddressError::InvalidPeerId(hash)
     }
 }