+++ src/transport/manager/mod.rs
@@ -42,7 +42,6 @@ use address::{scores, AddressStore};
 use futures::{future::BoxFuture, stream::FuturesUnordered, Stream, StreamExt};
 use indexmap::IndexMap;
 use multiaddr::{Multiaddr, Protocol};
-use multihash::Multihash;
 use parking_lot::RwLock;
 use tokio::sync::mpsc::{channel, Receiver, Sender};
 
@@ -485,9 +484,7 @@ impl TransportManager {
         let mut listen_addresses = self.listen_addresses.write();
 
         listen_addresses.insert(address.clone());
-        listen_addresses.insert(address.with(Protocol::P2p(
-            Multihash::from_bytes(&self.local_peer_id.to_bytes()).unwrap(),
-        )));
+        listen_addresses.insert(address.with(Protocol::P2p(self.local_peer_id.into())));
     }
 
     /// Add one or more known addresses for `peer`.
@@ -1468,8 +1465,6 @@ mod tests {
     use crate::transport::manager::{address::AddressStore, peer_state::SecondaryOrDialing};
     use limits::ConnectionLimitsConfig;
 
-    use multihash::Multihash;
-
     use super::*;
     use crate::{
         crypto::ed25519::Keypair,
@@ -1488,9 +1483,7 @@ mod tests {
         let dial_address = Multiaddr::empty()
             .with(Protocol::Ip4(Ipv4Addr::new(127, 0, 0, 1)))
             .with(Protocol::Tcp(8888 + connection_id))
-            .with(Protocol::P2p(
-                Multihash::from_bytes(&peer.to_bytes()).unwrap(),
-            ));
+            .with(Protocol::P2p(peer.into()));
         let connection_id = ConnectionId::from(connection_id as usize);
 
         (dial_address, connection_id)
@@ -1796,9 +1789,7 @@ mod tests {
             .with(Protocol::Ip4(Ipv4Addr::new(127, 0, 0, 1)))
             .with(Protocol::Udp(8888))
             .with(Protocol::QuicV1)
-            .with(Protocol::P2p(
-                Multihash::from_bytes(&PeerId::random().to_bytes()).unwrap(),
-            ));
+            .with(Protocol::P2p(PeerId::random().into()));
 
         assert!(std::matches!(
             manager.dial_address(address).await,
@@ -1817,9 +1808,7 @@ mod tests {
         let dial_address = Multiaddr::empty()
             .with(Protocol::Ip4(Ipv4Addr::new(127, 0, 0, 1)))
             .with(Protocol::Tcp(8888))
-            .with(Protocol::P2p(
-                Multihash::from_bytes(&peer.to_bytes()).unwrap(),
-            ));
+            .with(Protocol::P2p(peer.into()));
 
         let transport = Box::new({
             let mut transport = DummyTransport::new();
@@ -1876,9 +1865,7 @@ mod tests {
         let dial_address = Multiaddr::empty()
             .with(Protocol::Ip4(Ipv4Addr::new(127, 0, 0, 1)))
             .with(Protocol::Tcp(8888))
-            .with(Protocol::P2p(
-                Multihash::from_bytes(&peer.to_bytes()).unwrap(),
-            ));
+            .with(Protocol::P2p(peer.into()));
 
         assert!(manager.dial_address(dial_address.clone()).await.is_ok());
         assert_eq!(manager.pending_connections.len(), 1);
@@ -1904,9 +1891,7 @@ mod tests {
                 Multiaddr::empty()
                     .with(Protocol::Ip4(Ipv4Addr::new(127, 0, 0, 1)))
                     .with(Protocol::Tcp(8888))
-                    .with(Protocol::P2p(
-                        Multihash::from_bytes(&peer.to_bytes()).unwrap(),
-                    ))
+                    .with(Protocol::P2p(peer.into(),))
             )
             .await
             .is_ok());
@@ -1917,9 +1902,7 @@ mod tests {
                 Multiaddr::empty()
                     .with(Protocol::Ip6(Ipv6Addr::new(0, 0, 0, 0, 0, 0, 0, 1)))
                     .with(Protocol::Tcp(8888))
-                    .with(Protocol::P2p(
-                        Multihash::from_bytes(&peer.to_bytes()).unwrap(),
-                    ))
+                    .with(Protocol::P2p(peer.into(),))
             )
             .await
             .is_ok());
@@ -1980,18 +1963,14 @@ mod tests {
         let address = Multiaddr::empty()
             .with(Protocol::Ip6(Ipv6Addr::new(0, 0, 0, 0, 0, 0, 0, 1)))
             .with(Protocol::Tcp(8888))
-            .with(Protocol::P2p(
-                Multihash::from_bytes(&PeerId::random().to_bytes()).unwrap(),
-            ));
+            .with(Protocol::P2p(PeerId::random().into()));
         assert!(handle.supported_transport(&address));
 
         // ipv4
         let address = Multiaddr::empty()
             .with(Protocol::Ip4(Ipv4Addr::new(127, 0, 0, 1)))
             .with(Protocol::Tcp(8888))
-            .with(Protocol::P2p(
-                Multihash::from_bytes(&PeerId::random().to_bytes()).unwrap(),
-            ));
+            .with(Protocol::P2p(PeerId::random().into()));
         assert!(handle.supported_transport(&address));
 
         // quic
@@ -1999,9 +1978,7 @@ mod tests {
             .with(Protocol::Ip4(Ipv4Addr::new(127, 0, 0, 1)))
             .with(Protocol::Udp(8888))
             .with(Protocol::QuicV1)
-            .with(Protocol::P2p(
-                Multihash::from_bytes(&PeerId::random().to_bytes()).unwrap(),
-            ));
+            .with(Protocol::P2p(PeerId::random().into()));
         #[cfg(feature = "quic")]
         assert!(handle.supported_transport(&address));
         #[cfg(not(feature = "quic"))]
@@ -2038,15 +2015,11 @@ mod tests {
         let dial_address = Multiaddr::empty()
             .with(Protocol::Ip4(Ipv4Addr::new(127, 0, 0, 1)))
             .with(Protocol::Tcp(8888))
-            .with(Protocol::P2p(
-                Multihash::from_bytes(&peer.to_bytes()).unwrap(),
-            ));
+            .with(Protocol::P2p(peer.into()));
         let connect_address = Multiaddr::empty()
             .with(Protocol::Ip4(Ipv4Addr::new(192, 168, 1, 173)))
             .with(Protocol::Tcp(8888))
-            .with(Protocol::P2p(
-                Multihash::from_bytes(&peer.to_bytes()).unwrap(),
-            ));
+            .with(Protocol::P2p(peer.into()));
         assert!(manager.dial_address(dial_address.clone()).await.is_ok());
         assert_eq!(manager.pending_connections.len(), 1);
 
@@ -2099,15 +2072,11 @@ mod tests {
         let dial_address = Multiaddr::empty()
             .with(Protocol::Ip4(Ipv4Addr::new(127, 0, 0, 1)))
             .with(Protocol::Tcp(8888))
-            .with(Protocol::P2p(
-                Multihash::from_bytes(&peer.to_bytes()).unwrap(),
-            ));
+            .with(Protocol::P2p(peer.into()));
         let connect_address = Multiaddr::empty()
             .with(Protocol::Ip4(Ipv4Addr::new(192, 168, 1, 173)))
             .with(Protocol::Tcp(8888))
-            .with(Protocol::P2p(
-                Multihash::from_bytes(&peer.to_bytes()).unwrap(),
-            ));
+            .with(Protocol::P2p(peer.into()));
         assert!(manager.dial_address(dial_address.clone()).await.is_ok());
         assert_eq!(manager.pending_connections.len(), 1);
 
@@ -2180,15 +2149,11 @@ mod tests {
         let dial_address = Multiaddr::empty()
             .with(Protocol::Ip4(Ipv4Addr::new(127, 0, 0, 1)))
             .with(Protocol::Tcp(8888))
-            .with(Protocol::P2p(
-                Multihash::from_bytes(&peer.to_bytes()).unwrap(),
-            ));
+            .with(Protocol::P2p(peer.into()));
         let connect_address = Multiaddr::empty()
             .with(Protocol::Ip4(Ipv4Addr::new(192, 168, 1, 173)))
             .with(Protocol::Tcp(8888))
-            .with(Protocol::P2p(
-                Multihash::from_bytes(&peer.to_bytes()).unwrap(),
-            ));
+            .with(Protocol::P2p(peer.into()));
         assert!(manager.dial_address(dial_address.clone()).await.is_ok());
         assert_eq!(manager.pending_connections.len(), 1);
 
@@ -2258,21 +2223,15 @@ mod tests {
         let address1 = Multiaddr::empty()
             .with(Protocol::Ip4(Ipv4Addr::new(127, 0, 0, 1)))
             .with(Protocol::Tcp(8888))
-            .with(Protocol::P2p(
-                Multihash::from_bytes(&peer.to_bytes()).unwrap(),
-            ));
+            .with(Protocol::P2p(peer.into()));
         let address2 = Multiaddr::empty()
             .with(Protocol::Ip4(Ipv4Addr::new(192, 168, 1, 173)))
             .with(Protocol::Tcp(8888))
-            .with(Protocol::P2p(
-                Multihash::from_bytes(&peer.to_bytes()).unwrap(),
-            ));
+            .with(Protocol::P2p(peer.into()));
         let address3 = Multiaddr::empty()
             .with(Protocol::Ip4(Ipv4Addr::new(192, 168, 10, 64)))
             .with(Protocol::Tcp(9999))
-            .with(Protocol::P2p(
-                Multihash::from_bytes(&peer.to_bytes()).unwrap(),
-            ));
+            .with(Protocol::P2p(peer.into()));
 
         // remote peer connected to local node
         let established_result = manager
@@ -2361,15 +2320,11 @@ mod tests {
         let address1 = Multiaddr::empty()
             .with(Protocol::Ip4(Ipv4Addr::new(127, 0, 0, 1)))
             .with(Protocol::Tcp(8888))
-            .with(Protocol::P2p(
-                Multihash::from_bytes(&peer.to_bytes()).unwrap(),
-            ));
+            .with(Protocol::P2p(peer.into()));
         let address2 = Multiaddr::empty()
             .with(Protocol::Ip4(Ipv4Addr::new(192, 168, 1, 173)))
             .with(Protocol::Tcp(8888))
-            .with(Protocol::P2p(
-                Multihash::from_bytes(&peer.to_bytes()).unwrap(),
-            ));
+            .with(Protocol::P2p(peer.into()));
 
         // remote peer connected to local node
         let established_result = manager
@@ -2451,15 +2406,11 @@ mod tests {
         let address1 = Multiaddr::empty()
             .with(Protocol::Ip4(Ipv4Addr::new(127, 0, 0, 1)))
             .with(Protocol::Tcp(8888))
-            .with(Protocol::P2p(
-                Multihash::from_bytes(&peer.to_bytes()).unwrap(),
-            ));
+            .with(Protocol::P2p(peer.into()));
         let address2 = Multiaddr::empty()
             .with(Protocol::Ip4(Ipv4Addr::new(192, 168, 1, 173)))
             .with(Protocol::Tcp(8888))
-            .with(Protocol::P2p(
-                Multihash::from_bytes(&peer.to_bytes()).unwrap(),
-            ));
+            .with(Protocol::P2p(peer.into()));
 
         // remote peer connected to local node
         let emit_event = manager
@@ -2554,15 +2505,11 @@ mod tests {
         let address1 = Multiaddr::empty()
             .with(Protocol::Ip4(Ipv4Addr::new(127, 0, 0, 1)))
             .with(Protocol::Tcp(8888))
-            .with(Protocol::P2p(
-                Multihash::from_bytes(&peer.to_bytes()).unwrap(),
-            ));
+            .with(Protocol::P2p(peer.into()));
         let address2 = Multiaddr::empty()
             .with(Protocol::Ip4(Ipv4Addr::new(192, 168, 1, 173)))
             .with(Protocol::Tcp(8888))
-            .with(Protocol::P2p(
-                Multihash::from_bytes(&peer.to_bytes()).unwrap(),
-            ));
+            .with(Protocol::P2p(peer.into()));
 
         // remote peer connected to local node
         let emit_event = manager
@@ -2652,21 +2599,15 @@ mod tests {
         let address1 = Multiaddr::empty()
             .with(Protocol::Ip4(Ipv4Addr::new(127, 0, 0, 1)))
             .with(Protocol::Tcp(8888))
-            .with(Protocol::P2p(
-                Multihash::from_bytes(&peer.to_bytes()).unwrap(),
-            ));
+            .with(Protocol::P2p(peer.into()));
         let address2 = Multiaddr::empty()
             .with(Protocol::Ip4(Ipv4Addr::new(192, 168, 1, 173)))
             .with(Protocol::Tcp(8888))
-            .with(Protocol::P2p(
-                Multihash::from_bytes(&peer.to_bytes()).unwrap(),
-            ));
+            .with(Protocol::P2p(peer.into()));
         let address3 = Multiaddr::empty()
             .with(Protocol::Ip4(Ipv4Addr::new(192, 168, 1, 173)))
             .with(Protocol::Tcp(9999))
-            .with(Protocol::P2p(
-                Multihash::from_bytes(&peer.to_bytes()).unwrap(),
-            ));
+            .with(Protocol::P2p(peer.into()));
 
         // remote peer connected to local node
         let emit_event = manager
@@ -2915,7 +2856,7 @@ mod tests {
                             address: Multiaddr::empty()
                                 .with(Protocol::Ip4(std::net::Ipv4Addr::new(127, 0, 0, 1)))
                                 .with(Protocol::Tcp(8888))
-                                .with(Protocol::P2p(Multihash::from(peer))),
+                                .with(Protocol::P2p(peer.into())),
                             connection_id: ConnectionId::from(0usize),
                         },
                         secondary: None,
@@ -2925,7 +2866,7 @@ mod tests {
                         vec![Multiaddr::empty()
                             .with(Protocol::Ip4(std::net::Ipv4Addr::new(127, 0, 0, 1)))
                             .with(Protocol::Tcp(8888))
-                            .with(Protocol::P2p(Multihash::from(peer)))]
+                            .with(Protocol::P2p(peer.into()))]
                         .into_iter(),
                     ),
                 },
@@ -2957,7 +2898,7 @@ mod tests {
                             address: Multiaddr::empty()
                                 .with(Protocol::Ip4(std::net::Ipv4Addr::new(127, 0, 0, 1)))
                                 .with(Protocol::Tcp(8888))
-                                .with(Protocol::P2p(Multihash::from(peer))),
+                                .with(Protocol::P2p(peer.into())),
                             connection_id: ConnectionId::from(0usize),
                         },
                     },
@@ -2966,7 +2907,7 @@ mod tests {
                         vec![Multiaddr::empty()
                             .with(Protocol::Ip4(std::net::Ipv4Addr::new(127, 0, 0, 1)))
                             .with(Protocol::Tcp(8888))
-                            .with(Protocol::P2p(Multihash::from(peer)))]
+                            .with(Protocol::P2p(peer.into()))]
                         .into_iter(),
                     ),
                 },
@@ -2990,7 +2931,7 @@ mod tests {
                         Multiaddr::empty()
                             .with(Protocol::Ip4(std::net::Ipv4Addr::new(127, 0, 0, 1)))
                             .with(Protocol::Tcp(8888))
-                            .with(Protocol::P2p(Multihash::from(peer)))
+                            .with(Protocol::P2p(peer.into()))
                     );
                 }
                 state => panic!("invalid state: {state:?}"),
@@ -3015,7 +2956,7 @@ mod tests {
                             Multiaddr::empty()
                                 .with(Protocol::Ip4(std::net::Ipv4Addr::new(127, 0, 0, 1)))
                                 .with(Protocol::Tcp(8888))
-                                .with(Protocol::P2p(Multihash::from(peer))),
+                                .with(Protocol::P2p(peer.into())),
                             ConnectionId::from(0),
                         )),
                     },
@@ -3041,7 +2982,7 @@ mod tests {
 
         // transport doesn't start with ip/dns
         {
-            let address = Multiaddr::empty().with(Protocol::P2p(Multihash::from(PeerId::random())));
+            let address = Multiaddr::empty().with(Protocol::P2p(PeerId::random().into()));
             match manager.dial_address(address.clone()).await {
                 Err(Error::TransportNotSupported(dial_address)) => {
                     assert_eq!(dial_address, address);
@@ -3056,7 +2997,7 @@ mod tests {
                 .with(Protocol::Ip4(std::net::Ipv4Addr::new(127, 0, 0, 1)))
                 .with(Protocol::Udp(8888))
                 .with(Protocol::Utp)
-                .with(Protocol::P2p(Multihash::from(PeerId::random())));
+                .with(Protocol::P2p(PeerId::random().into()));
             match manager.dial_address(address.clone()).await {
                 Err(Error::TransportNotSupported(dial_address)) => {
                     assert_eq!(dial_address, address);
@@ -3070,7 +3011,7 @@ mod tests {
             let address = Multiaddr::empty()
                 .with(Protocol::Ip4(std::net::Ipv4Addr::new(127, 0, 0, 1)))
                 .with(Protocol::Sctp(8888))
-                .with(Protocol::P2p(Multihash::from(PeerId::random())));
+                .with(Protocol::P2p(PeerId::random().into()));
             match manager.dial_address(address.clone()).await {
                 Err(Error::TransportNotSupported(dial_address)) => {
                     assert_eq!(dial_address, address);
@@ -3085,7 +3026,7 @@ mod tests {
                 .with(Protocol::Ip4(std::net::Ipv4Addr::new(127, 0, 0, 1)))
                 .with(Protocol::Tcp(8888))
                 .with(Protocol::Utp)
-                .with(Protocol::P2p(Multihash::from(PeerId::random())));
+                .with(Protocol::P2p(PeerId::random().into()));
             match manager.dial_address(address.clone()).await {
                 Err(Error::TransportNotSupported(dial_address)) => {
                     assert_eq!(dial_address, address);
@@ -3150,9 +3091,7 @@ mod tests {
         let dial_address = Multiaddr::empty()
             .with(Protocol::Ip4(Ipv4Addr::new(127, 0, 0, 1)))
             .with(Protocol::Tcp(8888))
-            .with(Protocol::P2p(
-                Multihash::from_bytes(&peer.to_bytes()).unwrap(),
-            ));
+            .with(Protocol::P2p(peer.into()));
 
         let connection_id = ConnectionId::random();
         let transport = Box::new({
@@ -3169,7 +3108,7 @@ mod tests {
             vec![Multiaddr::empty()
                 .with(Protocol::Ip4(Ipv4Addr::new(192, 168, 1, 5)))
                 .with(Protocol::Tcp(8888))
-                .with(Protocol::P2p(Multihash::from(peer)))]
+                .with(Protocol::P2p(peer.into()))]
             .into_iter(),
         );
 
@@ -3230,9 +3169,7 @@ mod tests {
         let dial_address = Multiaddr::empty()
             .with(Protocol::Ip4(Ipv4Addr::new(127, 0, 0, 1)))
             .with(Protocol::Tcp(8888))
-            .with(Protocol::P2p(
-                Multihash::from_bytes(&peer.to_bytes()).unwrap(),
-            ));
+            .with(Protocol::P2p(peer.into()));
 
         let connection_id = ConnectionId::random();
         let transport = Box::new({
@@ -3249,7 +3186,7 @@ mod tests {
             vec![Multiaddr::empty()
                 .with(Protocol::Ip4(Ipv4Addr::new(127, 0, 0, 1)))
                 .with(Protocol::Tcp(8888))
-                .with(Protocol::P2p(Multihash::from(peer)))]
+                .with(Protocol::P2p(peer.into()))]
             .into_iter(),
         );
 
@@ -3525,9 +3462,7 @@ mod tests {
             let dial_address = Multiaddr::empty()
                 .with(Protocol::Ip4(Ipv4Addr::new(127, 0, 0, 1)))
                 .with(Protocol::Tcp(8888 + connection_id))
-                .with(Protocol::P2p(
-                    Multihash::from_bytes(&peer.to_bytes()).unwrap(),
-                ));
+                .with(Protocol::P2p(peer.into()));
             let connection_id = ConnectionId::from(connection_id as usize);
 
             (dial_address, connection_id)
@@ -3666,9 +3601,7 @@ mod tests {
         let dial_address = Multiaddr::empty()
             .with(Protocol::Ip4(Ipv4Addr::new(127, 0, 0, 1)))
             .with(Protocol::Tcp(8888))
-            .with(Protocol::P2p(
-                Multihash::from_bytes(&peer.to_bytes()).unwrap(),
-            ));
+            .with(Protocol::P2p(peer.into()));
 
         let connection_id = ConnectionId::from(0);
         let transport = Box::new({
@@ -3704,9 +3637,7 @@ mod tests {
         let second_address = Multiaddr::empty()
             .with(Protocol::Ip4(Ipv4Addr::new(127, 0, 0, 1)))
             .with(Protocol::Tcp(8889))
-            .with(Protocol::P2p(
-                Multihash::from_bytes(&peer.to_bytes()).unwrap(),
-            ));
+            .with(Protocol::P2p(peer.into()));
 
         // Second dial attempt with different address.
         manager.dial_address(second_address.clone()).await.unwrap();
@@ -3749,9 +3680,7 @@ mod tests {
         let dial_address_tcp = Multiaddr::empty()
             .with(Protocol::Ip4(Ipv4Addr::new(127, 0, 0, 1)))
             .with(Protocol::Tcp(8888))
-            .with(Protocol::P2p(
-                Multihash::from_bytes(&peer.to_bytes()).unwrap(),
-            ));
+            .with(Protocol::P2p(peer.into()));
         let transport = Box::new({
             let mut transport = DummyTransport::new();
             transport.inject_event(TransportEvent::OpenFailure {
@@ -3766,7 +3695,7 @@ mod tests {
             vec![Multiaddr::empty()
                 .with(Protocol::Ip4(Ipv4Addr::new(192, 168, 1, 5)))
                 .with(Protocol::Tcp(8888))
-                .with(Protocol::P2p(Multihash::from(peer)))]
+                .with(Protocol::P2p(peer.into()))]
             .into_iter(),
         );
 
@@ -3775,9 +3704,7 @@ mod tests {
             .with(Protocol::Ip4(Ipv4Addr::new(127, 0, 0, 1)))
             .with(Protocol::Tcp(8889))
             .with(Protocol::Ws(Cow::Borrowed("/")))
-            .with(Protocol::P2p(
-                Multihash::from_bytes(&peer.to_bytes()).unwrap(),
-            ));
+            .with(Protocol::P2p(peer.into()));
 
         let transport = Box::new({
             let mut transport = DummyTransport::new();
@@ -3794,9 +3721,7 @@ mod tests {
                 .with(Protocol::Ip4(Ipv4Addr::new(192, 168, 1, 5)))
                 .with(Protocol::Tcp(8889))
                 .with(Protocol::Ws(Cow::Borrowed("/")))
-                .with(Protocol::P2p(
-                    Multihash::from_bytes(&peer.to_bytes()).unwrap(),
-                ))]
+                .with(Protocol::P2p(peer.into()))]
             .into_iter(),
         );